Replace uint8 indexes by uint256

+++ contracts/core/MocBaseBucket.sol
@@ -256,7 +256,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param qTP_ amount of Pegged Token to add
      * @param qAC_ amount of Collateral Asset to add
      */
-    function _depositTP(uint8 i_, uint256 qTP_, uint256 qAC_) internal {
+    function _depositTP(uint256 i_, uint256 qTP_, uint256 qAC_) internal {
         pegContainer[i_].nTP += qTP_;
         _depositAC(qAC_);
     }
@@ -267,7 +267,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param qTP_ amount of Pegged Token to subtract
      * @param qAC_ amount of Collateral Asset to subtract
      */
-    function _withdrawTP(uint8 i_, uint256 qTP_, uint256 qAC_) internal {
+    function _withdrawTP(uint256 i_, uint256 qTP_, uint256 qAC_) internal {
         pegContainer[i_].nTP -= qTP_;
         nACcb -= qAC_;
     }
@@ -279,7 +279,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param qAC_ amount of Collateral Asset to add
      * @param recipient_ the account to mint tokens to
      */
-    function _depositAndMintTP(uint8 i_, uint256 qTP_, uint256 qAC_, address recipient_) internal {
+    function _depositAndMintTP(uint256 i_, uint256 qTP_, uint256 qAC_, address recipient_) internal {
         // add qTP and qAC to the Bucket
         _depositTP(i_, qTP_, qAC_);
         // mint qTP to the recipient
@@ -293,7 +293,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param qAC_ amount of Collateral Asset to subtract
      * @param toBurnFrom_ the account to burn tokens from
      */
-    function _withdrawAndBurnTP(uint8 i_, uint256 qTP_, uint256 qAC_, address toBurnFrom_) internal {
+    function _withdrawAndBurnTP(uint256 i_, uint256 qTP_, uint256 qAC_, address toBurnFrom_) internal {
         // sub qTP and qAC from the Bucket
         _withdrawTP(i_, qTP_, qAC_);
         // burn qTP from this address
@@ -411,12 +411,12 @@ abstract contract MocBaseBucket is MocUpgradable {
         uint256[] memory pACtps = new uint256[](pegAmount);
         // for each peg, calculates the proportion of AC reserves it's locked
 
-        for (uint8 i = 0; i < pegAmount; i = unchecked_inc(i)) {
+        for (uint256 i = 0; i < pegAmount; i = unchecked_inc(i)) {
             pACtps[i] = getPACtp(i);
             // [N] = [N] * [PREC] / [PREC]
             lckAC += _divPrec(pegContainer[i].nTP, pACtps[i]);
         }
-        for (uint8 i = 0; i < pegAmount; i = unchecked_inc(i)) {
+        for (uint256 i = 0; i < pegAmount; i = unchecked_inc(i)) {
             // [PREC] = [PREC] * [N] / [N];
             tpLiqPrices.push((pACtps[i] * lckAC) / totalACAvailable);
         }
@@ -427,7 +427,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param i_ Pegged Token index
      * @param pACtp_ Pegged Token price [PREC]
      */
-    function _updateTPtracking(uint8 i_, uint256 pACtp_) internal {
+    function _updateTPtracking(uint256 i_, uint256 pACtp_) internal {
         tpiou[i_] += _calcOtfPnLTP(i_, pACtp_);
         pACtpLstop[i_] = pACtp_;
     }
@@ -438,7 +438,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param pACtp_ Pegged Token price [PREC]
      * @return otfPnLtp [N]
      */
-    function _calcOtfPnLTP(uint8 i_, uint256 pACtp_) internal view returns (int256 otfPnLtp) {
+    function _calcOtfPnLTP(uint256 i_, uint256 pACtp_) internal view returns (int256 otfPnLtp) {
         // [PREC] = [N] * [PREC]
         uint256 nTP = pegContainer[i_].nTP * PRECISION;
         // [N] = [PREC] / [PREC] - [PREC] / [PREC]
@@ -452,7 +452,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @return tpGain amount of Pegged Token to be minted during settlement [N]
      * @return adjPnLtpi total amount of P&L in Collateral Asset [N]
      */
-    function _getPnLTP(uint8 i_, uint256 pACtp_) internal view returns (uint256 tpGain, uint256 adjPnLtpi) {
+    function _getPnLTP(uint256 i_, uint256 pACtp_) internal view returns (uint256 tpGain, uint256 adjPnLtpi) {
         // [N] = [N] + [N]
         int256 adjPnLtpiAux = tpiou[i_] + _calcOtfPnLTP(i_, pACtp_);
         if (adjPnLtpiAux > 0) {
@@ -471,7 +471,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      */
     function _getLckACandACgain() internal view returns (uint256 lckAC, uint256 nACgain) {
         uint256 pegAmount = pegContainer.length;
-        for (uint8 i = 0; i < pegAmount; i = unchecked_inc(i)) {
+        for (uint256 i = 0; i < pegAmount; i = unchecked_inc(i)) {
             uint256 pACtp = getPACtp(i);
             (uint256 tpGain, uint256 adjPnLtpi) = _getPnLTP(i, pACtp);
             // [N] = ([N] + [N]) * [PREC] / [PREC]
@@ -563,7 +563,7 @@ abstract contract MocBaseBucket is MocUpgradable {
      * @param i_ Pegged Token index
      * @return price [PREC]
      */
-    function getPACtp(uint8 i_) public view virtual returns (uint256) {
+    function getPACtp(uint256 i_) public view virtual returns (uint256) {
         IPriceProvider priceProvider = pegContainer[i_].priceProvider;
         (bytes32 price, bool has) = priceProvider.peek();
         if (!has) revert InvalidPriceProvider(address(priceProvider));
